+++ crates/module-system/sov-modules-macros/tests/integration/rpc_gen/rpc_gen_associated_types_nested.rs
@@ -40,6 +40,7 @@ pub trait Data:
     + UniversalWallet
     + borsh::BorshDeserialize
     + schemars::JsonSchema
+    + UniversalWallet
     + Send
     + Sync
     + 'static